requirements.txt
asgiref==3.5.2
Django==4.0.4
image==1.5.33
Pillow==9.1.1
PyYAML==6.0
six==1.16.0
sqlparse==0.4.2
